+++ src/main/java/org/betonquest/betonquest/api/feature/FeatureAPI.java
@@ -9,11 +9,13 @@ import org.betonquest.betonquest.feature.QuestCompass;
 import org.betonquest.betonquest.feature.journal.JournalMainPageEntry;
 import org.betonquest.betonquest.id.CompassID;
 import org.betonquest.betonquest.id.ConversationID;
+import org.betonquest.betonquest.id.ItemID;
 import org.betonquest.betonquest.id.JournalEntryID;
 import org.betonquest.betonquest.id.JournalMainPageID;
 import org.betonquest.betonquest.id.NpcID;
 import org.betonquest.betonquest.id.QuestCancelerID;
 import org.betonquest.betonquest.instruction.variable.VariableIdentifier;
+import org.betonquest.betonquest.item.QuestItem;
 import org.betonquest.betonquest.kernel.processor.QuestRegistry;
 import org.betonquest.betonquest.message.ParsedSectionMessage;
 
@@ -138,5 +140,16 @@ public final class FeatureAPI {
      */
     public NpcHider getNpcHider() {
         return questRegistry.npcs().getNpcHider();
+    }
+
+    /**
+     * Gets a QuestItem by their id.
+     *
+     * @param itemID the id
+     * @return the stored quest item
+     * @throws QuestException if there exists no QuestItem with that id
+     */
+    public QuestItem getItem(final ItemID itemID) throws QuestException {
+        return questRegistry.items().get(itemID);
     }
 }