+++ src/main/java/org/betonquest/betonquest/feature/Backpack.java
@@ -14,7 +14,6 @@ import org.betonquest.betonquest.database.PlayerData;
 import org.betonquest.betonquest.feature.journal.Journal;
 import org.betonquest.betonquest.id.CompassID;
 import org.betonquest.betonquest.id.ItemID;
-import org.betonquest.betonquest.item.QuestItem;
 import org.betonquest.betonquest.util.Utils;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
@@ -309,7 +308,8 @@ public class Backpack implements Listener {
                 present = true;
                 if (!checkDefault || !"DEFAULT".equalsIgnoreCase(buttonString)) {
                     try {
-                        stack = new QuestItem(new ItemID(null, buttonString)).generate(1);
+                        final ItemID itemId = new ItemID(null, buttonString);
+                        stack = BetonQuest.getInstance().getFeatureAPI().getItem(itemId).generate(1);
                     } catch (final QuestException e) {
                         log.warn("Could not load " + button + " button: " + e.getMessage(), e);
                     }
@@ -536,7 +536,7 @@ public class Backpack implements Listener {
                 }
                 ItemStack compass;
                 try {
-                    compass = new QuestItem(item).generate(1);
+                    compass = BetonQuest.getInstance().getFeatureAPI().getItem(item).generate(1);
                 } catch (final QuestException e) {
                     log.warn("Could not find item: " + e.getMessage(), e);
                     compass = new ItemStack(Material.COMPASS);