+++ src/main/java/org/betonquest/betonquest/menu/RPGMenu.java
@@ -4,6 +4,7 @@ import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.LanguageProvider;
 import org.betonquest.betonquest.api.config.ConfigAccessor;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
+import org.betonquest.betonquest.api.feature.FeatureAPI;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
@@ -59,6 +60,11 @@ public class RPGMenu {
     private final QuestTypeAPI questTypeAPI;
 
     /**
+     * Feature API.
+     */
+    private final FeatureAPI featureAPI;
+
+    /**
      * The profile provider instance.
      */
     private final ProfileProvider profileProvider;
@@ -74,12 +80,13 @@ public class RPGMenu {
 
     public RPGMenu(final BetonQuestLogger log, final BetonQuestLoggerFactory loggerFactory, final ConfigAccessor config,
                    final PluginMessage pluginMessage, final QuestTypeAPI questTypeAPI,
-                   final ProfileProvider profileProvider, final LanguageProvider languageProvider) {
+                   final FeatureAPI featureAPI, final ProfileProvider profileProvider, final LanguageProvider languageProvider) {
         this.log = log;
         this.loggerFactory = loggerFactory;
         this.config = config;
         this.pluginMessage = pluginMessage;
         this.questTypeAPI = questTypeAPI;
+        this.featureAPI = featureAPI;
         this.profileProvider = profileProvider;
         this.languageProvider = languageProvider;
         this.menus = new HashMap<>();
@@ -182,7 +189,7 @@ public class RPGMenu {
                 try {
                     final MenuID menuID = new MenuID(pack, name);
                     this.menus.put(menuID, new Menu(loggerFactory.create(Menu.class), loggerFactory, this, config,
-                            pluginMessage, questTypeAPI, profileProvider, menuID, languageProvider));
+                            pluginMessage, questTypeAPI, featureAPI, profileProvider, menuID, languageProvider));
                     info.loaded++;
                 } catch (final InvalidConfigurationException e) {
                     log.warn(pack, e.getMessage());
@@ -215,7 +222,7 @@ public class RPGMenu {
         final ReloadInformation info = new ReloadInformation();
         try {
             this.menus.put(menuID, new Menu(loggerFactory.create(Menu.class), loggerFactory, this, config,
-                    pluginMessage, questTypeAPI, profileProvider, menuID, languageProvider));
+                    pluginMessage, questTypeAPI, featureAPI, profileProvider, menuID, languageProvider));
             info.result = ReloadResult.FULL_SUCCESS;
             info.loaded = 1;
             log.info(menuID.getPackage(), "Reloaded menu " + menuID);