+++ src/main/java/org/betonquest/betonquest/quest/objective/brew/BrewObjectiveFactory.java
@@ -1,6 +1,8 @@
 package org.betonquest.betonquest.quest.objective.brew;
 
 import org.betonquest.betonquest.api.Objective;
+import org.betonquest.betonquest.api.logger.BetonQuestLogger;
+import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.profile.ProfileProvider;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.objective.ObjectiveFactory;
@@ -13,6 +15,12 @@ import org.betonquest.betonquest.instruction.variable.VariableNumber;
  * Factory for creating {@link BrewObjective} instances from {@link Instruction}s.
  */
 public class BrewObjectiveFactory implements ObjectiveFactory {
+
+    /**
+     * Logger factory to create a logger for the objectives.
+     */
+    private final BetonQuestLoggerFactory loggerFactory;
+
     /**
      * Profile provider to get the profile of the player.
      */
@@ -21,9 +29,11 @@ public class BrewObjectiveFactory implements ObjectiveFactory {
     /**
      * Creates a new instance of the BrewObjectiveFactory.
      *
+     * @param loggerFactory   the logger factory to create a logger for the objectives
      * @param profileProvider the profile provider to get the profile of the player
      */
-    public BrewObjectiveFactory(final ProfileProvider profileProvider) {
+    public BrewObjectiveFactory(final BetonQuestLoggerFactory loggerFactory, final ProfileProvider profileProvider) {
+        this.loggerFactory = loggerFactory;
         this.profileProvider = profileProvider;
     }
 
@@ -31,6 +41,7 @@ public class BrewObjectiveFactory implements ObjectiveFactory {
     public Objective parseInstruction(final Instruction instruction) throws QuestException {
         final Item potion = instruction.getItem();
         final VariableNumber targetAmount = instruction.get(VariableArgument.NUMBER_NOT_LESS_THAN_ZERO);
-        return new BrewObjective(instruction, targetAmount, profileProvider, potion);
+        final BetonQuestLogger log = loggerFactory.create(BrewObjective.class);
+        return new BrewObjective(instruction, targetAmount, log, profileProvider, potion);
     }
 }