+++ src/main/java/org/betonquest/betonquest/quest/objective/crafting/CraftingObjectiveFactory.java
@@ -1,6 +1,8 @@
 package org.betonquest.betonquest.quest.objective.crafting;
 
 import org.betonquest.betonquest.api.Objective;
+import org.betonquest.betonquest.api.logger.BetonQuestLogger;
+import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.objective.ObjectiveFactory;
 import org.betonquest.betonquest.instruction.Instruction;
@@ -12,16 +14,26 @@ import org.betonquest.betonquest.instruction.variable.VariableNumber;
  * Factory for creating {@link CraftingObjective} instances from {@link Instruction}s.
  */
 public class CraftingObjectiveFactory implements ObjectiveFactory {
+
+    /**
+     * Logger factory to create a logger for the objectives.
+     */
+    private final BetonQuestLoggerFactory loggerFactory;
+
     /**
      * Creates a new CraftingObjectiveFactory instance.
+     *
+     * @param loggerFactory the logger factory to create a logger for the objectives
      */
-    public CraftingObjectiveFactory() {
+    public CraftingObjectiveFactory(final BetonQuestLoggerFactory loggerFactory) {
+        this.loggerFactory = loggerFactory;
     }
 
     @Override
     public Objective parseInstruction(final Instruction instruction) throws QuestException {
         final Item item = instruction.getItem();
         final VariableNumber targetAmount = instruction.get(VariableArgument.NUMBER_NOT_LESS_THAN_ONE);
-        return new CraftingObjective(instruction, targetAmount, item);
+        final BetonQuestLogger log = loggerFactory.create(CraftingObjective.class);
+        return new CraftingObjective(instruction, targetAmount, log, item);
     }
 }