+++ src/main/java/org/betonquest/betonquest/quest/objective/equip/EquipItemObjectiveFactory.java
@@ -2,6 +2,8 @@ package org.betonquest.betonquest.quest.objective.equip;
 
 import com.destroystokyo.paper.event.player.PlayerArmorChangeEvent;
 import org.betonquest.betonquest.api.Objective;
+import org.betonquest.betonquest.api.logger.BetonQuestLogger;
+import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.objective.ObjectiveFactory;
 import org.betonquest.betonquest.instruction.Instruction;
@@ -12,15 +14,24 @@ import org.betonquest.betonquest.instruction.Item;
  */
 public class EquipItemObjectiveFactory implements ObjectiveFactory {
     /**
-     * Creates a new instance of the EquipItemObjectiveFactory.
+     * Logger factory to create a logger for the objectives.
      */
-    public EquipItemObjectiveFactory() {
+    private final BetonQuestLoggerFactory loggerFactory;
+
+    /**
+     * Creates a new instance of the EquipItemObjectiveFactory.
+     *
+     * @param loggerFactory the logger factory to create a logger for the objectives
+     */
+    public EquipItemObjectiveFactory(final BetonQuestLoggerFactory loggerFactory) {
+        this.loggerFactory = loggerFactory;
     }
 
     @Override
     public Objective parseInstruction(final Instruction instruction) throws QuestException {
         final PlayerArmorChangeEvent.SlotType slotType = instruction.getEnum(PlayerArmorChangeEvent.SlotType.class);
         final Item item = instruction.getItem();
-        return new EquipItemObjective(instruction, item, slotType);
+        final BetonQuestLogger log = loggerFactory.create(EquipItemObjective.class);
+        return new EquipItemObjective(instruction, log, item, slotType);
     }
 }